requirements.txt
requests==2.31.0
anthropic==0.25.0
pyyaml==6.0.1
click==8.1.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
feedparser==6.0.10